+++ src/sim-firmware.h
@@ -30,7 +30,7 @@ public:
   virtual void WaitQueueEmpty() {}
   virtual void MotorEnable(bool on) {}
   virtual void Shutdown(bool flush_queue) {}
-  virtual unsigned int GetPendingElements(uint32_t *head_item_progress) {
+  virtual size_t GetPendingElements(uint32_t *head_item_progress) {
     return 0;
   }
